Move `http.android.kt`
